+++ src/app/components/shared/resources/types.ts
@@ -1,0 +1,10 @@
+export interface Resource {
+  id: string;
+  name: string;
+  type: "file" | "link" | "image" | "text" | "directory" | "code";
+  source?: string;
+  url?: string;
+  content?: string;
+  fileSize?: number;
+  directoryPath?: string;
+}
